feat: move version model
