+++ pomice/models/__init__.py
@@ -1,0 +1,6 @@
+import pydantic
+from pydantic import ConfigDict
+
+
+class BaseModel(pydantic.BaseModel):
+    model_config = ConfigDict(populate_by_name=True)
